requirements.txt
beautifulsoup4==4.6.0
boto==2.48.0
boto3==1.7.48
botocore==1.10.48
bs4==0.0.1
bz2file==0.98
certifi==2018.4.16
chardet==3.0.4
cymem==1.31.2
cytoolz==0.8.2
dill==0.2.8.2
dj-database-url==0.4.2
Django==2.0.1
django-cors-headers==2.1.0
docutils==0.14
en-core-web-md==2.0.0
en-core-web-sm==2.0.0
ftfy==4.4.3
futures==3.1.1
gensim==3.4.0
googlemaps==3.0.2
goslate==1.5.1
gunicorn==19.7.1
html2text==2018.1.9
html5lib==1.0.1
idna==2.7
inflect==0.3.1
jaraco.itertools==2.3
Jinja2==2.10
jmespath==0.9.3
lxml==4.1.1
MarkupSafe==1.0
more-itertools==4.2.0
msgpack==0.5.4
msgpack-numpy==0.4.1
msgpack-python==0.5.6
murmurhash==0.28.0
nltk==3.3
numpy==1.14.5
pandas==0.22.0
pathlib==1.0.1
plac==0.9.6
preshed==1.0.0
psycopg2==2.7.3.2
pyodbc==4.0.23
pyshorteners==0.6.1
python-dateutil==2.7.3
python-decouple==3.1
python-Levenshtein==0.12.0
pytz==2017.3
pyxDamerauLevenshtein==1.5
regex==2017.4.5
requests==2.19.1
s3transfer==0.1.13
scikit-learn==0.19.1
scipy==1.1.0
six==1.11.0
slackclient==1.2.1
smart-open==1.6.0
spacy==2.0.5
tabulate==0.8.2
termcolor==1.1.0
textblob==0.15.1
thinc==6.10.2
toolz==0.9.0
tqdm==4.19.5
ujson==1.35
unicode==2.6
Unidecode==1.0.22
urllib3==1.23
virtualenv==15.1.0
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.48.0
whitenoise==3.3.1
wolframalpha==3.0.1
wrapt==1.10.11
xmltodict==0.11.0
python-forecastio==1.4.0
textblob==0.15.1
wolframalpha==3.0.1
pyshorteners==0.6.1
